refactor(abw): split research profile modules
